move CommonEstimateMemoryVisitor to runtime-common and enable it for K2
