+++ compiler/code-gen/declarations.cpp
@@ -908,9 +908,7 @@ void ClassDeclaration::compile_accept_json_visitor(CodeGenerator &W, ClassPtr kl
 
 void ClassDeclaration::compile_accept_visitor_methods(CodeGenerator &W, ClassPtr klass) {
   bool need_generic_accept =
-    klass->need_to_array_debug_visitor ||
-    (klass->need_instance_cache_visitors && !G->is_output_mode_k2()) ||
-    (klass->need_instance_memory_estimate_visitor && !G->is_output_mode_k2());
+    klass->need_to_array_debug_visitor || (klass->need_instance_cache_visitors && !G->is_output_mode_k2()) || (klass->need_instance_memory_estimate_visitor);
 
   if (!need_generic_accept && klass->json_encoders.empty()) {
     return;
@@ -926,10 +924,9 @@ void ClassDeclaration::compile_accept_visitor_methods(CodeGenerator &W, ClassPtr
     compile_accept_visitor(W, klass, "ToArrayVisitor");
   }
 
-  if ((klass->need_instance_memory_estimate_visitor ||
-       // for kphp_instance_cache_value_size statshouse metrics
-       klass->need_instance_cache_visitors)
-      && !G->is_output_mode_k2()) {
+  if (klass->need_instance_memory_estimate_visitor ||
+      // for kphp_instance_cache_value_size statshouse metrics
+      (klass->need_instance_cache_visitors && !G->is_output_mode_k2())) {
     W << NL;
     compile_accept_visitor(W, klass, "CommonMemoryEstimateVisitor");
   }
@@ -1063,9 +1060,7 @@ void ClassDeclaration::compile_job_worker_shared_memory_piece_methods(CodeGenera
 
 void ClassMembersDefinition::compile(CodeGenerator &W) const {
   bool need_generic_accept =
-    klass->need_to_array_debug_visitor ||
-    (klass->need_instance_cache_visitors && !G->is_output_mode_k2()) ||
-    (klass->need_instance_memory_estimate_visitor && !G->is_output_mode_k2());
+    klass->need_to_array_debug_visitor || (klass->need_instance_cache_visitors && !G->is_output_mode_k2()) || (klass->need_instance_memory_estimate_visitor);
 
   if (!need_generic_accept && !klass->is_serializable && klass->json_encoders.empty()) {
     return;
@@ -1090,10 +1085,9 @@ void ClassMembersDefinition::compile(CodeGenerator &W) const {
     compile_generic_accept_instantiations(W, klass, "ToArrayVisitor");
   }
 
-  if ((klass->need_instance_memory_estimate_visitor ||
-       // for kphp_instance_cache_value_size statshouse metrics
-       klass->need_instance_cache_visitors)
-      && !G->is_output_mode_k2()) {
+  if (klass->need_instance_memory_estimate_visitor ||
+      // for kphp_instance_cache_value_size statshouse metrics
+      (klass->need_instance_cache_visitors && !G->is_output_mode_k2())) {
     W << NL;
     compile_generic_accept_instantiations(W, klass, "CommonMemoryEstimateVisitor");
   }
