+++ runtime-light/stdlib/system/system-functions.h
@@ -11,11 +11,6 @@
 #include "runtime-light/state/image-state.h"
 #include "runtime-light/stdlib/system/system-state.h"
 
-template<typename T>
-int64_t f$estimate_memory_usage(const T & /*unused*/) {
-  php_critical_error("call to unsupported function");
-}
-
 template<typename F>
 void f$register_kphp_on_warning_callback(F && /*callback*/) {
   php_critical_error("call to unsupported function");
