+++ runtime/exception.h
@@ -11,8 +11,8 @@
 #include "runtime-common/core/class-instance/refcountable-php-classes.h"
 #include "runtime-common/core/runtime-core.h"
 #include "runtime-common/stdlib/visitors/dummy-visitor-methods.h"
+#include "runtime-common/stdlib/visitors/memory-visitors.h"
 #include "runtime/instance-copy-processor.h"
-#include "runtime/memory_usage.h"
 #include "runtime/to-array-processor.h"
 
 array<array<string>> f$debug_backtrace();
