+++ tests/phpt/memory_usage/10_arrays.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 
 function test_vector() {
