+++ tests/phpt/memory_usage/11_shapes.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 require_once 'kphp_tester_include.php';
