+++ tests/phpt/memory_usage/1_primitive_types.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 
 function test_primitive_types() {
