+++ tests/phpt/memory_usage/3_lambdas.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 
 function test_empty_lambda() {
