+++ tests/phpt/memory_usage/4_interfaces.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 
 interface IMyClass {
