+++ tests/phpt/memory_usage/5_polymorphic_classes.php
@@ -1,4 +1,4 @@
-@ok k2_skip
+@ok
 <?php
 
 class A {
